+++ src/test/java/de/jfachwert/formular/AnredeTest.java
@@ -1,5 +1,5 @@
 package de.jfachwert.formular;/*
- * Copyright (c) 2017 by Oliver Boehm
+ * Copyright (c) 2017-2024 by Oliver Boehm
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -68,12 +68,23 @@ public final class AnredeTest extends FachwertTest {
         assertEquals(Anrede.VEREINIGUNG, Anrede.of(6));
     }
 
+
+    @Test
+    public void testOfDamen() {
+        assertEquals(Anrede.DAMEN, Anrede.of(7));
+    }
+
+    @Test
+    public void testOfHerren() {
+        assertEquals(Anrede.HERREN, Anrede.of(8));
+    }
+
     /**
      * Test-Methode fuer {@link Anrede#of(int)} mit fehlerhaftem Argument.
      */
     @Test
     public void testOfInvalidArgument() {
-        assertThrows(ValidationException.class, () -> Anrede.of(7));
+        assertThrows(ValidationException.class, () -> Anrede.of(9));
     }
 
 }